I was forced to introduce CommandType.
Initially the commands were supposed to be polymorphic and have a virtual void Execute function. This would allow them to be self-contained. At some point they became a ledger.
However, I failed to predict that I would need RTTI. This is a problem because they are plain C++ classes and Cast<T> can only be used on UObjects. I can't use dynamic_cast since it's disabled. This situation forced me to introduce a CommandType. A refactor is not an option due to time constraints.